updating modal to have tabs
