+++ client/src/components/CodeJar/index.css
@@ -21,6 +21,7 @@
   height: 385px;
   width: 100%;
   overflow: scroll;
+  overflow-x: hidden;
   /* overflow-wrap: normal; */
   /* border-radius: 10px;  */
 }
